requirements.txt
# Core dependencies with specific versions
typer==0.9.0
rich==13.9.4
click==8.1.8
shellingham==1.5.4

# Database and security
python-dotenv==1.0.1
sentry-sdk[sqlalchemy]==1.40.0
SQLAlchemy==2.0.25
bcrypt==4.2.1

# Testing and development
Faker==22.5.0

# Other required dependencies
certifi==2025.1.31
ecdsa==0.19.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pyasn1==0.4.8
Pygments==2.19.1
python-jose==3.4.0
rsa==4.9
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
